+++ package.json
@@ -16,7 +16,6 @@
     "prettier": "^2.2.1"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
